+++ lib/reconcile/vernemq.go
@@ -468,7 +468,7 @@ func getVerneMQPolicyRules() []rbacv1.PolicyRule {
 	return []rbacv1.PolicyRule{
 		{
 			APIGroups: []string{""},
-			Resources: []string{"pods", "services"},
+			Resources: []string{"pods", "services", "endpoints"},
 			Verbs:     []string{"list", "get"},
 		},
 		{